Stack idea (hopefully) prepared, but not implemented.
Pulling references from throughout TeX.SX (mostly egreg, TBH) and the
documentation for l3seq (via `texdoc interface3`), I've abstracted the
idea of a breadcrumb into a stack so that it is presentation-agnostic.

This will allow the package to be much more extensible; as long as I
can get the stack to behave correctly from page to page, the
presentation doesn't matter.  (How's that for an idealized LaTeX
concept for you?)

I've made a function that will (hopefully; untested) build the
sequence into the input stream delimited by the separator defined as
`\l_breadcrumb_sep_tl`, set via l3keys with `\BreadcrumbSet`.

I've also included a yet-unused 'font' key to control the font the
breadcrumb will use when it comes time to throw it into TeX's stomach,
or whatever the correct terminology is.  (The library's copy of The
TeXbook is coming to my campus within the next few days.  I still
haven't read it, but this problem is about to be fixed.)